Increase configuration maximum of keyboard auto repeat rate
The current maximum configurable keyboard auto-repeat rate of 100 is too conservative. Change it to 200.

For me personally, repeat of about 130/s is ideal. Increasing much beyond 130 is too fast for me (too hard to move cursor short distances). I set the maximum to 200 on the assumption that there are people with much faster reflexes than me but 200 "ought to be enough for anybody".